refactor: Transitioning the session save process
